フォルダ階層の変更、namespaceの追加
