--- clicker_game/Assets/Clicker/Components/Item/Class/UpgradeComponent.cs
@@ -1,42 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public enum UpgradeComponentEnum
-{
-    Money,
-    Component,
-}
-
-public class UpgradeComponent
-{
-    private UpgradeComponentEnum upgradeComponentType;
-    public UpgradeComponentEnum UpgradeComponentType
-    {
-        get
-        {
-            return upgradeComponentType;
-        }
-    }
-
-    private int count;
-    public int Count
-    {
-        get
-        {
-            return count;
-        }
-    }
-
-    public UpgradeComponent(UpgradeComponentEnum upgradeComponentType, int count)
-    {
-        this.upgradeComponentType = upgradeComponentType;
-        this.count = count;
-    }
-
-    public void AddCount(int addCount)
-    {
-        count += addCount;
-    }
-}
