+++ clicker_game/Assets/Clicker/Components/Item/UpgradeComponent.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 023a0e1002837ff40b2d9e18bc59f2c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
