+++ clicker_game/Assets/Clicker/Components/Item/UpgradeableItem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 458ed9c9d3513574b92eb3bb5cc4f52e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
