+++ clicker_game/Assets/Clicker/Components/Item/UpgradeableItem/Class.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c473b26e9edd3f347a89353af5dc86a9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
