+++ clicker_game/Assets/Clicker/Components/Item/UpgradeableItem/Class/UpgradeableItem.cs.meta
