+++ clicker_game/Assets/Clicker/Components/Item/UpgradeableItem/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ffea6b95fb707204ca55209a77826770
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
