+++ clicker_game/Assets/Clicker/Components/Item/UpgradeableItem/Prefabs/UpgradeableItemPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bfc64f08e0ad2214ba16036af7c05670
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
