+++ clicker_game/Assets/Clicker/Components/Item/UpgradeableItem/View.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dfd5d337b810eec4fbeb4be8427a188d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
